added bool cut for truth match selection in streams
